Create the comment page
In this change, I create a navbar component with navigating pages and I also create the a simle scroll of comments in the commenting page
