+++ package.json
@@ -9,6 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.45.1",
+    "@tanstack/react-query-devtools": "^5.45.1",
+    "@tanstack/react-query-next-experimental": "^5.45.1",
     "autoprefixer": "^10.4.19",
     "next": "14.2.3",
     "postcss": "^8.4.38",
