+++ pages/users.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Users() {
+  return <div>Users</div>;
+}
+
+export default Users;
